requirements.txt
djangorestframework==3.14.0
Django==4.2.3
black==22.10.0
drf-yasg==1.21.7
celery==5.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.6.0
